Create INSTALL.target documentation
This document explains how to setup a target to receive "twist" and "pretzel" commands from a host "twist-host.some.domain.com" with IP address 192.168.1.100 using that host's host principal keytab.  A service account on the twist host could be used instead, but the service account and k5login would need to be created on the target.